Fix : module delete
